implement Special:ProtectPage

+++ ottm/api/auth.py
@@ -250,9 +250,12 @@ def block_user(user: models.User, performer: models.User, allow_messages_on_own_
     :param end_date: The date until which to block this user. None means infinite.
     :param reason: The block’s reason.
     :raise MissingPermissionError: If the performer does not have the "block_users" permission.
+    :raise PastDateError: If the date is in the past.
     """
     if performer and not performer.has_permission(_perms.PERM_BLOCK_USERS):
         raise errors.MissingPermissionError(_perms.PERM_BLOCK_USERS)
+    if end_date and end_date <= end_date.now().date():
+        raise errors.PastDateError()
     for block in models.UserBlock.objects.filter(user=user.internal_object):
         block.delete()
     models.UserBlock(
